pyproject.toml
[tool.poetry]
name = "ml-tech-assessment"
version = "0.1.0"
description = ""
authors = ["jhonvalderrama <jhonvalderrama@aceup.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.76.2"
pydantic-settings = "^2.9.1"
pytest = "^8.3.5"
fastapi = {extras = ["standard"], version = "^0.128.0"}
redis = {extras = ["hiredis"], version = "^7.1.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-cov (>=7.0.0,<8.0.0)"
]
